connect.cpp: fix warnings: comparison of integer expressions of different signednes

+++ bundled_deps/admesh/admesh/connect.cpp
@@ -482,9 +482,10 @@ void stl_check_facets_nearby(stl_file *stl, float tolerance)
 {
 	assert(stl->stats.connected_facets_3_edge <= stl->stats.connected_facets_2_edge);
 	assert(stl->stats.connected_facets_2_edge <= stl->stats.connected_facets_1_edge);
-	assert(stl->stats.connected_facets_1_edge <= stl->stats.number_of_facets);
+	// cast number_of_facets to int to avoid signed/unsigned comparison warnings
+	assert(stl->stats.connected_facets_1_edge <= static_cast<int>(stl->stats.number_of_facets));
 
-  	if (stl->stats.connected_facets_3_edge == stl->stats.number_of_facets)
+  	if (stl->stats.connected_facets_3_edge == static_cast<int>(stl->stats.number_of_facets))
     	// No need to check any further.  All facets are connected.
     	return;
 
@@ -533,7 +534,8 @@ void stl_remove_unconnected_facets(stl_file *stl)
 		  	for (int i = 0; i < 3; ++ i)
 		    	if (neighbors.neighbor[i] != -1) {
 			    	int &other_face_idx = stl->neighbors_start[neighbors.neighbor[i]].neighbor[(neighbors.which_vertex_not[i] + 1) % 3];
-			  		if (other_face_idx != stl->stats.number_of_facets) {
+			  		// compare with casted value to avoid signed/unsigned mismatch
+			  		if (other_face_idx != static_cast<int>(stl->stats.number_of_facets)) {
 			  			BOOST_LOG_TRIVIAL(info) << "in remove_facet: neighbor = " << other_face_idx << " numfacets = " << stl->stats.number_of_facets << " this is wrong";
 			    		return;
 			  		}
